+++ lib/array/isNotEmpty.ts
@@ -1,9 +1,12 @@
+'use strict';
+
 /**
  * Check whether or not a provided value is an array with content
  *
  * @param val - Value to verify
- * 
+ *
  * @returns Whether or not the value is an array with content
  */
-declare function isNotEmptyArray(val:any):boolean;
-export = isNotEmptyArray;
+export default function isNotEmptyArray (val:any):boolean {
+    return Array.isArray(val) && val.length !== 0;
+}