+++ CHANGELOG.md
@@ -16,6 +16,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Dev Dep: Upgrade c8 to 9.1.0
 - function/sleep: Will now internally check whether or not the value passed to it is a number above 0 and fallback to 0 if that is not the case
 - deep/set: Will now reject paths containing \_\_proto\_\_, constructor, prototype as keys to prevent possibility of prototype pollution
+- array/mapPrimitive: Now allows passing a number for valround, which in turn will be used to round to a certain decimal precision
 
 ### Removed
 - Dev Dep: @babel/cli